Rename review > reviewPage
In order for it to be consistent with the other pages in this folder
