+++ cypress_shared/pages/assess/index.ts
@@ -1,7 +1,7 @@
 /* eslint-disable import/prefer-default-export */
 import ListPage from './listPage'
 import TaskListPage from './taskListPage'
-import ReviewPage from './review'
+import ReviewPage from './reviewPage'
 import SufficientInformationPage from './sufficientInformationPage'
 
 export { ListPage, TaskListPage, ReviewPage, SufficientInformationPage }
